fixes offscreen sprties issues

+++ LuaImplementation.h
@@ -87,8 +87,9 @@ static int draw_to_oam(lua_State* L)
 	ServerRAM.RAM[0x200 + oam_index] = tile;
 	ServerRAM.RAM[0x201 + oam_index] = size;
 
-	uint_fast16_t sprite_x_position = uint_fast16_t(int(offset_x + ServerRAM.RAM[0x2100 + sprite_index] + ServerRAM.RAM[0x2180 + sprite_index] * 256));
-	uint_fast16_t sprite_y_position = uint_fast16_t(int(offset_y + ServerRAM.RAM[0x2280 + sprite_index] + ServerRAM.RAM[0x2300 + sprite_index] * 256));
+
+	uint_fast16_t sprite_x_position = uint_fast16_t(int(offset_x + ServerRAM.RAM[0x2100 + sprite_index] + int_fast8_t(ServerRAM.RAM[0x2180 + sprite_index]) * 256));
+	uint_fast16_t sprite_y_position = uint_fast16_t(int(offset_y + ServerRAM.RAM[0x2280 + sprite_index] + int_fast8_t(ServerRAM.RAM[0x2300 + sprite_index]) * 256));
 
 
 	ServerRAM.RAM[0x202 + oam_index] = sprite_x_position;
